LPS-66213 Move class into portal-upgrade module
