+++ modules/apps/foundation/portal/portal-upgrade/bnd.bnd
@@ -1,7 +1,9 @@
 Bundle-Name: Liferay Portal Upgrade
 Bundle-SymbolicName: com.liferay.portal.upgrade
 Bundle-Version: 2.1.0
-Export-Package: com.liferay.portal.upgrade.registry
+Export-Package:\
+	com.liferay.portal.upgrade.legacy,\
+	com.liferay.portal.upgrade.registry
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Upgrade and Verify
 -includeresource:\
